[WIP] Simplifying code

+++ ft_printf/srcs/ft_printf/ft_printf_atoi.c
@@ -6,16 +6,16 @@
 /*   By: bopopovi <bopopovi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/12 16:18:15 by bopopovi          #+#    #+#             */
-/*   Updated: 2018/07/14 17:45:19 by bopopovi         ###   ########.fr       */
+/*   Updated: 2018/08/11 02:25:31 by bopopovi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_printf_atoi(const char *str, int *res)
+size_t		ft_printf_atoi(const char *str, int *res)
 {
 	int			flag;
-	int			i;
+	size_t		i;
 
 	i = 0;
 	*res = 0;
